Move existing API checkin and checkout tests under domain directory
